+++ server/package.json
@@ -5,7 +5,6 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "dev": "nodemon --watch ./ --ext ts --exec ts-node ./index.ts",
-    "ws" : "nodemon --watch ./ --ext ts --exec ts-node ./dev.ts",
     "build": "tsc"
   },
   "keywords": [],
@@ -13,9 +12,7 @@
   "license": "ISC",
   "description": "",
   "dependencies": {
-    "cors": "^2.8.5",
     "dotenv": "^17.2.0",
-    "express": "^5.1.0",
     "websocket": "^1.0.35",
     "ws": "^8.18.3"
   },
